+++ packages/qwik/src/core/shared/component-execution.ts
@@ -1,8 +1,8 @@
 import { isDev } from '@qwik.dev/core/build';
 import { vnode_isVNode } from '../client/vnode';
 import { Slot } from '../shared/jsx/slot.public';
-import { isSignal } from '../signal/signal';
-import { clearAllEffects } from '../signal/signal-cleanup';
+import { isSignal } from '../reactive-primitives/signal';
+import { clearAllEffects } from '../reactive-primitives/signal-cleanup';
 import { invokeApply, newInvokeContext, untrack } from '../use/use-core';
 import { type EventQRL, type UseOnMap } from '../use/use-on';
 import { isQwikComponent, type OnRenderFn } from './component.public';
@@ -25,8 +25,8 @@ import {
 } from './utils/markers';
 import { MAX_RETRY_ON_PROMISE_COUNT, isPromise, maybeThen, safeCall } from './utils/promises';
 import type { ValueOrPromise } from './utils/types';
-import { getSubscriber } from '../signal/subscriber';
-import { EffectProperty } from '../signal/types';
+import { getSubscriber } from '../reactive-primitives/subscriber';
+import { EffectProperty } from '../reactive-primitives/types';
 
 /**
  * Use `executeComponent` to execute a component.
