+++ packages/qwik/src/core/shared/jsx/types/jsx-qwik-attributes.ts
@@ -1,5 +1,5 @@
 import type { QRL } from '../../qrl/qrl.public';
-import type { Signal } from '../../../signal/signal.public';
+import type { Signal } from '../../../reactive-primitives/signal.public';
 import type { JSXNode } from './jsx-node';
 import type {
   QwikIdleEvent,
