+++ packages/qwik/src/core/shared/utils/serialize-utils.ts
@@ -3,8 +3,8 @@ import { isNode } from './element';
 import { isPromise } from './promises';
 import { isArray, isFunction, isObject, isSerializableObject } from './types';
 import { canSerialize } from '../shared-serialization';
-import { isSignal } from '../../signal/signal';
-import { unwrapStore } from '../../signal/store';
+import { isSignal } from '../../reactive-primitives/signal';
+import { unwrapStore } from '../../reactive-primitives/store';
 
 /** @internal */
 export const verifySerializable = <T>(value: T, preMessage?: string): T => {
