+++ packages/qwik/src/core/use/use-computed.ts
@@ -1,9 +1,9 @@
 import { implicit$FirstArg } from '../shared/qrl/implicit_dollar';
 import { assertQrl } from '../shared/qrl/qrl-utils';
 import type { QRL } from '../shared/qrl/qrl.public';
-import { ComputedSignalImpl } from '../signal/impl/computed-signal-impl';
-import { throwIfQRLNotResolved } from '../signal/signal';
-import type { ReadonlySignal, Signal } from '../signal/signal.public';
+import { ComputedSignalImpl } from '../reactive-primitives/impl/computed-signal-impl';
+import { throwIfQRLNotResolved } from '../reactive-primitives/signal';
+import type { ReadonlySignal, Signal } from '../reactive-primitives/signal.public';
 import { useSequentialScope } from './use-sequential-scope';
 
 /** @public */
