+++ packages/qwik/src/core/use/use-signal.ts
@@ -1,6 +1,6 @@
 import { isQwikComponent } from '../shared/component.public';
 import { isFunction } from '../shared/utils/types';
-import { createSignal, type Signal } from '../signal/signal.public';
+import { createSignal, type Signal } from '../reactive-primitives/signal.public';
 import { invoke } from './use-core';
 import { useSequentialScope } from './use-sequential-scope';
