+++ packages/qwik/src/core/use/use-task.ts
@@ -10,16 +10,21 @@ import { TaskEvent } from '../shared/utils/markers';
 import { isPromise, safeCall } from '../shared/utils/promises';
 import { noSerialize, type NoSerialize } from '../shared/utils/serialize-utils';
 import { isFunction, type ValueOrPromise } from '../shared/utils/types';
-import { isSignal } from '../signal/signal';
-import { BackRef, clearAllEffects } from '../signal/signal-cleanup';
-import { type Signal } from '../signal/signal.public';
+import { isSignal } from '../reactive-primitives/signal';
+import { BackRef, clearAllEffects } from '../reactive-primitives/signal-cleanup';
+import { type Signal } from '../reactive-primitives/signal.public';
 import { invoke, newInvokeContext } from './use-core';
 import { useLexicalScope } from './use-lexical-scope.public';
 import type { ResourceReturnInternal } from './use-resource';
 import { useSequentialScope } from './use-sequential-scope';
-import { getSubscriber } from '../signal/subscriber';
-import { addStoreEffect, getStoreHandler, getStoreTarget, isStore } from '../signal/store';
-import { EffectProperty, STORE_ALL_PROPS } from '../signal/types';
+import { getSubscriber } from '../reactive-primitives/subscriber';
+import {
+  addStoreEffect,
+  getStoreHandler,
+  getStoreTarget,
+  isStore,
+} from '../reactive-primitives/store';
+import { EffectProperty, STORE_ALL_PROPS } from '../reactive-primitives/types';
 
 export const enum TaskFlags {
   VISIBLE_TASK = 1 << 0,
